feat: refactor MCP CLI and server structure
