+++ packages/mcp/package.json
@@ -5,14 +5,13 @@
   "type": "module",
   "bin": {
     "argent": "dist/cli.js",
-    "argent-mcp": "dist/index.js",
     "argent-simulator-server": "bin/simulator-server"
   },
   "scripts": {
     "build": "npm run build:mcp && npm run build:tools-bundle",
     "build:mcp": "rm -rf dist tsconfig.tsbuildinfo && tsc",
     "build:tools-bundle": "node scripts/bundle-tools.cjs",
-    "dev": "node --loader ts-node/esm src/index.ts",
+    "dev": "node --loader ts-node/esm src/cli.ts mcp",
     "test": "vitest run",
     "test:watch": "vitest",
     "observe": "node scripts/observe.cjs",
